calendar_api_creation/requirements.txt
Flask==2.0.1
google-api-python-client==2.39.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
Werkzeug==2.0.3
